Rename: 기존 repository 파일 legacy로 분류
