+++ src/main/java/com/project/doubleshop/domain/delivery/service/DeliveryDriverService.java
@@ -8,7 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.project.doubleshop.domain.common.Status;
 import com.project.doubleshop.domain.delivery.entity.legacy.DeliveryDriver;
-import com.project.doubleshop.domain.delivery.repository.DeliveryDriverRepository;
+import com.project.doubleshop.domain.delivery.repository.legacy.DeliveryDriverRepository;
 import com.project.doubleshop.domain.exception.NotFoundException;
 import com.project.doubleshop.web.common.StatusRequest;
 import com.project.doubleshop.web.config.support.Pageable;
