+++ src/test/java/com/project/doubleshop/domain/delivery/service/DeliveryDriverServiceTest.java
@@ -13,7 +13,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.project.doubleshop.domain.delivery.entity.legacy.DeliveryDriver;
-import com.project.doubleshop.domain.delivery.repository.DeliveryDriverRepository;
+import com.project.doubleshop.domain.delivery.repository.legacy.DeliveryDriverRepository;
 import com.project.doubleshop.domain.exception.NotFoundException;
 import com.project.doubleshop.web.config.support.Pageable;
