+++ src/test/java/com/project/doubleshop/domain/delivery/service/DeliveryPolicyServiceTest.java
@@ -13,10 +13,9 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.project.doubleshop.domain.delivery.entity.DeliveryPolicy;
-import com.project.doubleshop.domain.delivery.repository.DeliveryPolicyRepository;
+import com.project.doubleshop.domain.delivery.repository.legacy.DeliveryPolicyRepository;
 import com.project.doubleshop.domain.exception.NotFoundException;
 import com.project.doubleshop.web.config.support.Pageable;
-import com.project.doubleshop.web.item.exception.DataNotFoundException;
 
 @ExtendWith(MockitoExtension.class)
 class DeliveryPolicyServiceTest {
