+++ src/test/java/com/project/doubleshop/domain/delivery/service/DeliveryServiceTest.java
@@ -13,10 +13,9 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.project.doubleshop.domain.delivery.entity.Delivery;
-import com.project.doubleshop.domain.delivery.repository.DeliveryRepository;
+import com.project.doubleshop.domain.delivery.repository.legacy.DeliveryRepository;
 import com.project.doubleshop.domain.exception.NotFoundException;
 import com.project.doubleshop.web.config.support.Pageable;
-import com.project.doubleshop.web.item.exception.DataNotFoundException;
 
 @ExtendWith(MockitoExtension.class)
 class DeliveryServiceTest {
